mercurius example -> graphql-ez
