+++ .npmrc
@@ -1,4 +1,4 @@
-workspace-concurrency=4
+workspace-concurrency=0
 public-hoist-pattern[]=@types/*
 public-hoist-pattern[]=esbuild
 public-hoist-pattern[]=typescript
