+++ examples/ez/gqty.config.cjs
