+++ examples/ez/jest.config.js
