+++ examples/ez/listen.ts
@@ -1,4 +1,4 @@
-const { app } = require('./');
+import { app } from './src/index';
 
 app.listen(3000, (err) => {
   if (err) {
