+++ examples/ez/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "example-mercurius",
+  "name": "example-ez",
   "version": "0.1.0",
   "private": true,
   "license": "MIT",
@@ -22,51 +22,39 @@
     "src"
   ],
   "scripts": {
-    "analyze": "size-limit --why",
     "build": "bob-esbuild build",
-    "generate": "ts-node --skip-project generate.ts",
-    "listen": "node listen.js",
+    "generate": "node -r esbuild-register generate.ts",
+    "listen": "node -r esbuild-register listen.ts",
     "prepare": "bob-esbuild build",
-    "size": "size-limit",
     "start": "bob-esbuild watch",
     "test": "jest",
     "test:watch": "bob-esbuild watch"
   },
   "dependencies": {
     "@gqty/cli": "workspace:^1.1.4",
+    "@graphql-ez/fastify": "^0.9.4",
+    "@graphql-ez/fastify-testing": "^0.1.2",
+    "@graphql-ez/plugin-altair": "^0.9.1",
+    "@graphql-ez/plugin-codegen": "^0.7.3",
+    "@graphql-ez/plugin-schema": "^0.7.3",
     "@graphql-typed-document-node/core": "^3.1.0",
     "fastify": "^3.21.1",
     "gqty": "workspace:^1.1.3",
     "graphql": "^15.5.3",
+    "graphql-ez": "^0.13.0",
     "lodash": "^4.17.21",
-    "mercurius": "^8.2.1",
-    "mercurius-codegen": "^3.0.0",
     "randomstring": "^1.2.1",
     "test-utils": "workspace:^0.1.0"
   },
   "devDependencies": {
-    "@size-limit/preset-small-lib": "^5.0.3",
     "@types/lodash": "^4.14.173",
     "@types/randomstring": "^1.1.7",
     "bob-esbuild-cli": "^1.0.1",
+    "esbuild-register": "^3.0.0",
     "jest": "^27.2.0",
-    "mercurius-integration-testing": "^3.2.0",
-    "size-limit": "^5.0.3",
-    "ts-node": "^10.2.1",
-    "tslib": "^2.3.1",
     "typescript": "^4.4.3"
   },
   "engines": {
     "node": ">=12"
-  },
-  "size-limit": [
-    {
-      "path": "dist/mercurius-example.cjs.production.min.js",
-      "limit": "10 KB"
-    },
-    {
-      "path": "dist/mercurius-example.esm.js",
-      "limit": "10 KB"
-    }
-  ]
+  }
 }
