+++ examples/ez/src/graphql/operations.gql
