+++ examples/ez/test/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig.json",
+  "include": ["example.test.ts"],
+  "exclude": []
+}
