+++ examples/ez/tsconfig.json
