--- examples/mercurius/.browserslistrc
@@ -1,1 +1,0 @@
-node 10
