+++ package.json
@@ -16,7 +16,6 @@
     "ci:version": "changeset version && pnpm i --no-frozen-lockfile --lockfile-only --ignore-scripts && pnpm update -r \"@gqty/*\" gqty",
     "clean": "rimraf packages/*/dist packages/*/coverage packages/*/.nyc_output packages/*/*.tsbuildinfo",
     "dev": "bob-esbuild tsc && pnpm -r dev --no-sort --no-bail --filter=!example-*",
-    "test:watch": "pnpm test:watch -r --no-bail --no-sort --stream",
     "docs:build": "pnpm build:website && pretty-quick",
     "prepare": "husky install",
     "pretty": "pretty-quick",
@@ -24,21 +23,10 @@
     "publish:all": "changeset version && pnpm i && pnpm test && pnpm -r publish --access public",
     "test": "pnpm test -r --no-sort --no-bail --stream --",
     "test:ci": "pnpm test -r --workspace-concurrency=0 --no-sort --no-bail --stream --",
+    "test:watch": "pnpm test:watch -r --no-bail --no-sort --stream",
     "typecheck": "bob-esbuild tsc",
     "website": "pnpm -r --filter website start",
     "website:build": "pnpm -r --filter website docs:build && pretty-quick"
-  },
-  "dependencies": {
-    "bufferutil": "^4.0.3",
-    "esbuild": "^0.12.28",
-    "jest": "^27.2.0",
-    "open": "^8.2.1",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.2.1",
-    "tslib": "^2.3.1",
-    "typescript": "^4.4.3",
-    "utf-8-validate": "^5.0.5",
-    "wait-on": "^6.0.0"
   },
   "devDependencies": {
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
@@ -48,16 +36,26 @@
     "bob-esbuild": "^1.3.0",
     "bob-esbuild-cli": "^1.0.1",
     "bob-ts": "^1.0.6",
+    "bufferutil": "^4.0.3",
     "chalk": "^4.1.2",
     "changesets-github-release": "^0.0.4",
     "concurrently": "^6.2.1",
+    "esbuild": "^0.12.28",
     "fkill": "^7.2.1",
     "globby": "^12.0.2",
     "graphql": "^15.5.3",
     "husky": "^7.0.2",
+    "jest": "^27.2.0",
+    "open": "^8.2.1",
     "prettier": "^2.4.1",
     "pretty-quick": "^3.1.1",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "ts-jest": "^27.0.5",
+    "ts-node": "^10.2.1",
+    "tslib": "^2.3.1",
+    "typescript": "^4.4.3",
+    "utf-8-validate": "^5.0.5",
+    "wait-on": "^6.0.0"
   },
   "engines": {
     "pnpm": ">=6.0.2"
